+++ lib/safedep/error.rb
@@ -1,0 +1,3 @@
+module Safedep
+  class Error < StandardError; end
+end